% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_algorithm_ext.R
\name{.extcomrank1}
\alias{.extcomrank1}
\title{Find all rank-1 communities under an extended DAG}
\usage{
EXTCOMRANK1(S,G)
}
\arguments{
\item{S}{A matrix whose columns are the elements to be covered, and whose rows are the available subsets from which to build the covers}

\item{G}{The adjacency matrix specifying the DAG, with added extension codes}

\item{verbose}{Controls level of output}
}
\value{
A list of all rank-1 communities
}
\description{
Find all rank-1 communities under an extended DAG
}
