% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_internals.R
\name{.getelders}
\alias{.getelders}
\title{From a community, get the indices representing elder nodes}
\usage{
.getelders(A, H)
}
\arguments{
\item{A}{A community}

\item{G}{the DAG}
}
\value{
Return the vector of elder nodes of A
}
\description{
From a community, get the indices representing elder nodes
}
