% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_internals.R
\name{.iscover}
\alias{.iscover}
\title{User utility to determine if a specified community is a valid cover}
\usage{
ISCOVER(tr, S)
}
\arguments{
\item{tr}{A community}

\item{S}{The subset family}
}
\value{
TRUE if tr is a cover of S;  FALSE otherwise
}
\description{
User utility to determine if a specified community is a valid cover
}
