% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_internals.R
\name{.isdag}
\alias{.isdag}
\title{Determines if given adjacency matrix represents a DAG}
\usage{
ISDAG(G)
}
\arguments{
\item{G}{An adjacency matrix}
}
\value{
TRUE if A represents a DAG;  FALSE otherwise
}
\description{
Determines if given adjacency matrix represents a DAG
}
