% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_internals.R
\name{.xflags}
\alias{.xflags}
\title{Returns the X-flags for a given community}
\usage{
.xflags(C, S)
}
\arguments{
\item{C}{A community}

\item{S}{The family of subsets}
}
\value{
Returns a table with members of X marked for each cover
}
\description{
Returns the X-flags for a given community
}
\examples{
XFLAGS(C,S)
}
