% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_interface.R
\name{dagmin}
\alias{dagmin}
\title{Find all minimally consistent covers of a set subject to the specified DAG constraint}
\usage{
dagmin(S, G, delimsep = ",", maxrank = NULL, verbose = TRUE)
}
\arguments{
\item{S}{A matrix whose columns are the elements to be covered, and whose rows are the available subsets from which to build the covers}

\item{G}{The adjacency matrix specifying the DAG}

\item{delimsep}{The delimeter for reading a text file}

\item{maxrank}{The maximum rank community to consider}

\item{verbose}{Controls level of output}
}
\value{
A list all minimally consistent covers and their s.flags and x.flags
}
\description{
Find all minimally consistent covers of a set subject to the specified DAG constraint
}
\examples{
dagmin(S,G)
}
