% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_algorithm_main.R
\name{.comrank1}
\alias{.comrank1}
\title{Internal routine to find all rank-1 communities}
\usage{
.comrank1(S, G)
}
\arguments{
\item{S}{A matrix whose columns are the elements to be covered, and whose rows are the available subsets from which to build the covers}

\item{G}{The adjacency matrix specifying the DAG}
}
\value{
A list of all rank-1 communities
}
\description{
Internal routine to find all rank-1 communities
}
\examples{
comrank1(S,G)
}
