% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_algorithm_main.R
\name{.comrankj}
\alias{.comrankj}
\title{Internal routine to find all rank-j communities, using a list of rank-1 communities as building blocks}
\usage{
.comrankj(j, D)
}
\arguments{
\item{j}{A positive integer specifying the desired rank}

\item{D}{A list of rank-1 communities as generated by COMRANK1}
}
\value{
A list of all rank-j communities
}
\description{
Internal routine to find all rank-j communities, using a list of rank-1 communities as building blocks
}
\examples{
COMRANKJ(S,G)
}
