% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_algorithm_ext.R
\name{.extended.G}
\alias{.extended.G}
\title{Creates a list of all possible partitions of an extended-choice DAG}
\usage{
.extended.G(G, verbose)
}
\arguments{
\item{G}{The adjacency matrix specifying the DAG, with added extension codes}

\item{verbose}{Controls level of output}
}
\value{
A list of all possible standard DAGs (one for each possible partition)
}
\description{
Creates a list of all possible partitions of an extended-choice DAG
}
\examples{
EXTENDED.G(G)
}
