% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_internals.R
\name{.getinput}
\alias{.getinput}
\title{Internal helper function for getting input data}
\usage{
.getinput(input, delimsep)
}
\arguments{
\item{input}{Either an existing matrix or a filename containing matrix in delimited text format}

\item{delimsep}{The delimeter for reading a text file}
}
\value{
The requested matrix
}
\description{
Internal helper function for getting input data
}
\examples{
getinput(input, delimsep)
}
