% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_algorithm_main.R
\name{.isdominant}
\alias{.isdominant}
\title{Determines if a specified community is dominant}
\usage{
.isdominant(t, S, P)
}
\arguments{
\item{t}{A community}

\item{S}{The subset family}

\item{P}{The path matrix}
}
\value{
TRUE if t is dominant;  FALSE otherwise
}
\description{
Determines if a specified community is dominant
}
\examples{
ISDOMINANT(t, S, P)
}
