% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_internals.R
\name{.issuperset}
\alias{.issuperset}
\title{Determines if one community is a superset of another}
\usage{
.issuperset(a, b)
}
\arguments{
\item{a}{A community, possibly a subset of b}

\item{b}{The reference community}
}
\value{
TRUE if a is a superset of b;  FALSE otherwise
}
\description{
Determines if one community is a superset of another
}
\examples{
ISSUPERSET(a, b)
}
