% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_internals.R
\name{.iswellformed}
\alias{.iswellformed}
\title{Determines if a specified community is "well-formed"}
\usage{
.iswellformed(tr, j, G, P)
}
\arguments{
\item{tr}{A community}

\item{j}{The rank of the community}

\item{G}{The DAG}

\item{P}{matrix P}
}
\value{
TRUE if tr is well-formed;  FALSE otherwise
}
\description{
Determines if a specified community is "well-formed"
}
\examples{
ISWELLFORMED(tr,j,G,P)
}
