% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_algorithm_main.R
\name{.mcc_enum}
\alias{.mcc_enum}
\title{Main internal routine for enumeration}
\usage{
.mcc_enum(S, G, verbose)
}
\arguments{
\item{S}{A matrix whose columns are the elements to be covered, and whose rows are the available subsets from which to build the covers}

\item{G}{The adjacency matrix specifying the DAG}
}
\value{
A list of all minimally consistent covers
}
\description{
Main internal routine for enumeration
}
\examples{
mcc_enum(S,G)
}
