% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_internals.R
\name{.sflags}
\alias{.sflags}
\title{Returns the S-flags for a given community}
\usage{
.sflags(C, S)
}
\arguments{
\item{C}{A community}

\item{S}{The family of subsets}
}
\value{
Returns a table with members of S marked for each cover
}
\description{
Returns the S-flags for a given community
}
\examples{
SFLAGS(C,S)
}
