% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmin_internals.R
\name{.validate}
\alias{.validate}
\title{Internal helper function for validating input data}
\usage{
.validate(S, G)
}
\arguments{
\item{S}{The family of subsets}

\item{G}{The adjacency matrix}
}
\value{
No return value.  Stops program if validation not successful.
}
\description{
Internal helper function for validating input data
}
\examples{
validate(S,G)
}
